refactor: move source files under a single root
