+++ app/__init__.py
@@ -3,7 +3,7 @@ from dotenv import load_dotenv
 
 
 # configure a logger
-logger = logging.getLogger("main")
+logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 # create console handler
@@ -16,5 +16,5 @@ ch.setFormatter(formatter)
 logger.addHandler(ch)
 
 
-# Load default environment variables. Will not override existing variables.
+# Load default environment variables. This will not override existing variables.
 load_dotenv(".env.dev")
