+++ app/static/img/heart.png
