+++ requirements.in
@@ -1,7 +1,7 @@
 beautifulsoup4
 flask
-google-cloud-storage
 google-cloud-secret-manager
+google-cloud-storage
 gunicorn
 jsonschema
 openai
