+++ requirements.txt
@@ -1,81 +1,181 @@
+#
+# This file is autogenerated by pip-compile with Python 3.10
+# by the following command:
+#
+#    pip-compile
+#
 annotated-types==0.7.0
+    # via pydantic
 anyio==4.4.0
-asttokens==2.0.5
+    # via
+    #   httpx
+    #   openai
 attrs==24.1.0
-backcall==0.2.0
+    # via
+    #   jsonschema
+    #   pytest
+    #   referencing
 beautifulsoup4==4.11.1
+    # via -r requirements.in
 blinker==1.8.2
+    # via flask
 cachetools==5.2.0
+    # via google-auth
 certifi==2022.6.15
+    # via
+    #   httpcore
+    #   httpx
+    #   requests
 charset-normalizer==2.1.0
+    # via requests
 click==8.1.3
-decorator==5.1.1
+    # via flask
 distro==1.9.0
+    # via openai
 exceptiongroup==1.2.2
-executing==0.8.3
-Flask==3.0.3
-google-api-core==2.19.1
+    # via anyio
+flask==3.0.3
+    # via -r requirements.in
+google-api-core[grpc]==2.19.1
+    # via
+    #   google-cloud-core
+    #   google-cloud-secret-manager
+    #   google-cloud-storage
 google-auth==2.33.0
+    # via
+    #   google-api-core
+    #   google-cloud-core
+    #   google-cloud-secret-manager
+    #   google-cloud-storage
 google-cloud-core==2.3.2
+    # via google-cloud-storage
 google-cloud-secret-manager==2.20.2
+    # via -r requirements.in
 google-cloud-storage==2.4.0
+    # via -r requirements.in
 google-crc32c==1.3.0
+    # via google-resumable-media
 google-resumable-media==2.3.3
-googleapis-common-protos==1.56.4
+    # via google-cloud-storage
+googleapis-common-protos[grpc]==1.56.4
+    # via
+    #   google-api-core
+    #   grpc-google-iam-v1
+    #   grpcio-status
 grpc-google-iam-v1==0.13.1
+    # via google-cloud-secret-manager
 grpcio==1.65.4
+    # via
+    #   google-api-core
+    #   googleapis-common-protos
+    #   grpc-google-iam-v1
+    #   grpcio-status
 grpcio-status==1.62.3
+    # via google-api-core
 gunicorn==20.1.0
+    # via -r requirements.in
 h11==0.14.0
+    # via httpcore
 httpcore==1.0.5
+    # via httpx
 httpx==0.27.0
+    # via openai
 idna==3.3
-importlib-metadata==7.0.1
+    # via
+    #   anyio
+    #   httpx
+    #   requests
 iniconfig==1.1.1
-ipython==8.4.0
+    # via pytest
 itsdangerous==2.1.2
-jedi==0.18.1
-Jinja2==3.1.2
+    # via flask
+jinja2==3.1.2
+    # via flask
 jiter==0.5.0
+    # via openai
 jsonschema==4.23.0
+    # via -r requirements.in
 jsonschema-specifications==2023.12.1
-MarkupSafe==2.1.1
-matplotlib-inline==0.1.3
+    # via jsonschema
+markupsafe==2.1.1
+    # via
+    #   jinja2
+    #   werkzeug
 openai==1.40.4
+    # via -r requirements.in
 packaging==21.3
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
+    # via pytest
 pillow==10.4.0
+    # via -r requirements.in
 pluggy==1.0.0
-prompt-toolkit==3.0.30
+    # via pytest
 proto-plus==1.24.0
+    # via
+    #   google-api-core
+    #   google-cloud-secret-manager
 protobuf==4.25.4
-ptyprocess==0.7.0
-pure-eval==0.2.2
+    # via
+    #   google-api-core
+    #   google-cloud-secret-manager
+    #   googleapis-common-protos
+    #   grpc-google-iam-v1
+    #   grpcio-status
+    #   proto-plus
 py==1.11.0
+    # via pytest
 pyasn1==0.4.8
+    # via
+    #   pyasn1-modules
+    #   rsa
 pyasn1-modules==0.2.8
+    # via google-auth
 pydantic==2.8.2
-pydantic_core==2.20.1
-Pygments==2.12.0
+    # via openai
+pydantic-core==2.20.1
+    # via pydantic
 pyparsing==3.0.9
+    # via packaging
 pytest==7.1.2
+    # via -r requirements.in
 python-dotenv==1.0.0
-PyYAML==6.0.2
+    # via -r requirements.in
+pyyaml==6.0.2
+    # via -r requirements.in
 referencing==0.35.1
+    # via
+    #   jsonschema
+    #   jsonschema-specifications
 requests==2.28.1
+    # via
+    #   google-api-core
+    #   google-cloud-storage
 rpds-py==0.19.1
+    # via
+    #   jsonschema
+    #   referencing
 rsa==4.9
-six==1.16.0
+    # via google-auth
 sniffio==1.3.1
+    # via
+    #   anyio
+    #   httpx
+    #   openai
 soupsieve==2.3.2.post1
-stack-data==0.3.0
+    # via beautifulsoup4
 tomli==2.0.1
+    # via pytest
 tqdm==4.66.5
-traitlets==5.3.0
-typing_extensions==4.12.2
+    # via openai
+typing-extensions==4.12.2
+    # via
+    #   anyio
+    #   openai
+    #   pydantic
+    #   pydantic-core
 urllib3==1.26.10
-wcwidth==0.2.5
-Werkzeug==3.0.3
-zipp==3.17.0
+    # via requests
+werkzeug==3.0.3
+    # via flask
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
